test : 피드백 수정에 맞게 테스트 코드 수정
